+++ src/contract_client/assembled_transaction.ts
@@ -1,30 +1,216 @@
 import type {
-  ContractClientOptions,
-  XDR_BASE64,
   AssembledTransactionOptions,
+  ContractClientOptions,
+  Tx,
+  XDR_BASE64,
 } from "./types";
 import {
   BASE_FEE,
   Contract,
-  Memo,
-  MemoType,
   Operation,
   SorobanRpc,
   StrKey,
-  Transaction,
   TransactionBuilder,
   authorizeEntry,
-  hash,
   xdr,
 } from "..";
 import { Err } from "../rust_types";
+import {
+  DEFAULT_TIMEOUT,
+  contractErrorPattern,
+  implementsToString
+} from "./utils";
+import { SentTransaction } from "./sent_transaction";
 
-const DEFAULT_TIMEOUT = 10;
-
-type Tx = Transaction<Memo<MemoType>, Operation[]>;
-
-type SendTx = SorobanRpc.Api.SendTransactionResponse;
-type GetTx = SorobanRpc.Api.GetTransactionResponse;
+/**
+ * The main workhorse of {@link ContractClient}. This class is used to wrap a
+ * transaction-under-construction and provide high-level interfaces to the most
+ * common workflows, while still providing access to low-level stellar-sdk
+ * transaction manipulation.
+ *
+ * Most of the time, you will not construct an `AssembledTransaction` directly,
+ * but instead receive one as the return value of a `ContractClient` method. If
+ * you're familiar with the libraries generated by soroban-cli's `contract
+ * bindings typescript` command, these also wraps `ContractClient` and return
+ * `AssembledTransaction` instances.
+ *
+ * Let's look at examples of how to use `AssembledTransaction` for a variety of
+ * use-cases:
+ *
+ * # 1. Simple read call
+ *
+ * Since these only require simulation, you can get the `result` of the call
+ * right after constructing your `AssembledTransaction`:
+ *
+ * ```ts
+ * const { result } = await AssembledTransaction.build({
+ *   method: 'myReadMethod',
+ *   args: spec.funcArgsToScVals('myReadMethod', { args: 'for', my: 'method', ... }),
+ *   contractId: 'C123…',
+ *   networkPassphrase: '…',
+ *   rpcUrl: 'https://…',
+ *   publicKey: Keypair.random().publicKey(), // keypairs are irrelevant, for simulation-only read calls
+ *   parseResultXdr: (result: xdr.ScVal) => spec.funcResToNative('myReadMethod', result),
+ * })
+ * ```
+ *
+ * While that looks pretty complicated, most of the time you will use this in
+ * conjunction with {@link ContractClient}, which simplifies it to:
+ *
+ * ```ts
+ * const { result }  = await client.myReadMethod({ args: 'for', my: 'method', ... })
+ * ```
+ *
+ * # 2. Simple write call
+ *
+ * For write calls that will be simulated and then sent to the network without
+ * further manipulation, only one more step is needed:
+ *
+ * ```ts
+ * const assembledTx = await client.myWriteMethod({ args: 'for', my: 'method', ... })
+ * const sentTx = await assembledTx.signAndSend()
+ * ```
+ *
+ * Here we're assuming that you're using a {@link ContractClient}, rather than
+ * constructing `AssembledTransaction`'s directly.
+ *
+ * Note that `sentTx`, the return value of `signAndSend`, is a
+ * {@link SentTransaction}. `SentTransaction` is similar to
+ * `AssembledTransaction`, but is missing many of the methods and fields that
+ * are only relevant while assembling a transaction. It also has a few extra
+ * methods and fields that are only relevant after the transaction has been
+ * sent to the network.
+ *
+ * Like `AssembledTransaction`, `SentTransaction` also has a `result` getter,
+ * which contains the parsed final return value of the contract call. Most of
+ * the time, you may only be interested in this, so rather than getting the
+ * whole `sentTx` you may just want to:
+ *
+ * ```ts
+ * const tx = await client.myWriteMethod({ args: 'for', my: 'method', ... })
+ * const { result } = await tx.signAndSend()
+ * ```
+ *
+ * # 3. More fine-grained control over transaction construction
+ *
+ * If you need more control over the transaction before simulating it, you can
+ * set various {@link MethodOptions} when constructing your
+ * `AssembledTransaction`. With a {@link ContractClient}, this is passed as a
+ * second object after the arguments (or the only object, if the method takes
+ * no arguments):
+ *
+ * ```ts
+ * const tx = await client.myWriteMethod(
+ *   {
+ *     args: 'for',
+ *     my: 'method',
+ *     ...
+ *   }, {
+ *     fee: '10000', // default: {@link BASE_FEE}
+ *     simulate: false,
+ *     timeoutInSeconds: 20, // default: {@link DEFAULT_TIMEOUT}
+ *   }
+ * )
+ * ```
+ *
+ * Since we've skipped simulation, we can now edit the `raw` transaction and
+ * then manually call `simulate`:
+ *
+ * ```ts
+ * tx.raw.addMemo(Memo.text('Nice memo, friend!'))
+ * await tx.simulate()
+ * ```
+ *
+ * If you need to inspect the simulation later, you can access it with `tx.simulation`.
+ *
+ * # 4. Multi-auth workflows
+ *
+ * Soroban, and Stellar in general, allows multiple parties to sign a
+ * transaction.
+ *
+ * Let's consider an Atomic Swap contract. Alice wants to give 10 of her Token
+ * A tokens to Bob for 5 of his Token B tokens.
+ *
+ * ```ts
+ * const ALICE = 'G123...'
+ * const BOB = 'G456...'
+ * const TOKEN_A = 'C123…'
+ * const TOKEN_B = 'C456…'
+ * const AMOUNT_A = 10n
+ * const AMOUNT_B = 5n
+ * ```
+ *
+ * Let's say Alice is also going to be the one signing the final transaction
+ * envelope, meaning she is the invoker. So your app, from Alice's browser,
+ * simulates the `swap` call:
+ *
+ * ```ts
+ * const tx = await swapClient.swap({
+ *   a: ALICE,
+ *   b: BOB,
+ *   token_a: TOKEN_A,
+ *   token_b: TOKEN_B,
+ *   amount_a: AMOUNT_A,
+ *   amount_b: AMOUNT_B,
+ * })
+ * ```
+ *
+ * But your app can't `signAndSend` this right away, because Bob needs to sign
+ * it first. You can check this:
+ *
+ * ```ts
+ * const whoElseNeedsToSign = tx.needsNonInvokerSigningBy()
+ * ```
+ *
+ * You can verify that `whoElseNeedsToSign` is an array of length `1`,
+ * containing only Bob's public key.
+ *
+ * Then, still on Alice's machine, you can serialize the
+ * transaction-under-assembly:
+ *
+ * ```ts
+ * const json = tx.toJSON()
+ * ```
+ *
+ * And now you need to send it to Bob's browser. How you do this depends on
+ * your app. Maybe you send it to a server first, maybe you use WebSockets, or
+ * maybe you have Alice text the JSON blob to Bob and have him paste it into
+ * your app in his browser (note: this option might be error-prone 😄).
+ *
+ * Once you get the JSON blob into your app on Bob's machine, you can
+ * deserialize it:
+ *
+ * ```ts
+ * const tx = swapClient.txFromJSON(json)
+ * ```
+ *
+ * Or, if you're using a client generated with `soroban contract bindings
+ * typescript`, this deserialization will look like:
+ *
+ * ```ts
+ * const tx = swapClient.fromJSON.swap(json)
+ * ```
+ *
+ * Then you can have Bob sign it. What Bob will actually need to sign is some
+ * _auth entries_ within the transaction, not the transaction itself or the
+ * transaction envelope. Your app can verify that Bob has the correct wallet
+ * selected, then:
+ *
+ * ```ts
+ * await tx.signAuthEntries()
+ * ```
+ *
+ * Under the hood, this uses `signAuthEntry`, which you either need to inject
+ * during initial construction of the `ContractClient`/`AssembledTransaction`,
+ * or which you can pass directly to `signAuthEntries`.
+ *
+ * Now Bob can again serialize the transaction and send back to Alice, where
+ * she can finally call `signAndSend()`.
+ *
+ * To see an even more complicated example, where Alice swaps with Bob but the
+ * transaction is invoked by yet another party, check out
+ * [test-swap.js](../../test/e2e/src/test-swap.js).
+ */
 export class AssembledTransaction<T> {
   /**
    * The TransactionBuilder as constructed in `{@link
@@ -57,14 +243,24 @@ export class AssembledTransaction<T> {
   public simulation?: SorobanRpc.Api.SimulateTransactionResponse;
   /**
    * Cached simulation result. This is set after the first call to
-   * `simulationData`, and is used to facilitate serialization and
+   * {@link simulationData}, and is used to facilitate serialization and
    * deserialization of the AssembledTransaction.
+   *
+   * Most of the time, if you need this data, you can call `tx.simulation.result`.
+   *
+   * If you need access to this data after a transaction has been serialized
+   * and then deserialized, you can call `simulationData.result`.
    */
   private simulationResult?: SorobanRpc.Api.SimulateHostFunctionResult;
   /**
    * Cached simulation transaction data. This is set after the first call to
-   * `simulationData`, and is used to facilitate serialization and
+   * {@link simulationData}, and is used to facilitate serialization and
    * deserialization of the AssembledTransaction.
+   *
+   * Most of the time, if you need this data, you can call `simulation.transactionData`.
+   *
+   * If you need access to this data after a transaction has been serialized
+   * and then deserialized, you can call `simulationData.transactionData`.
    */
   private simulationTransactionData?: xdr.SorobanTransactionData;
   /**
@@ -144,6 +340,24 @@ export class AssembledTransaction<T> {
     });
   }
 
+  /**
+   * Construct a new AssembledTransaction. This is the only way to create a new
+   * AssembledTransaction; the main constructor is private.
+   *
+   * This is an asynchronous constructor for two reasons:
+   *
+   * 1. It needs to fetch the account from the network to get the current
+   *   sequence number.
+   * 2. It needs to simulate the transaction to get the expected fee.
+   *
+   * If you don't want to simulate the transaction, you can set `simulate` to
+   * `false` in the options.
+   *
+   *     const tx = await AssembledTransaction.build({
+   *       ...,
+   *       simulate: false,
+   *     })
+   */
   static async build<T>(
     options: AssembledTransactionOptions<T>
   ): Promise<AssembledTransaction<T>> {
@@ -153,7 +367,7 @@ export class AssembledTransaction<T> {
     const account = await tx.server.getAccount(options.publicKey);
 
     tx.raw = new TransactionBuilder(account, {
-      fee: options.fee?.toString(10) ?? BASE_FEE,
+      fee: options.fee ?? BASE_FEE,
       networkPassphrase: options.networkPassphrase,
     })
       .addOperation(contract.call(options.method, ...(options.args ?? [])))
@@ -244,7 +458,7 @@ export class AssembledTransaction<T> {
     }
   }
 
-  parseError(errorMessage: string) {
+  private parseError(errorMessage: string) {
     if (!this.options.errorTypes) return undefined;
     const match = errorMessage.match(contractErrorPattern);
     if (!match) return undefined;
@@ -303,7 +517,7 @@ export class AssembledTransaction<T> {
     return await SentTransaction.init(signTransaction, typeChecked);
   };
 
-  getStorageExpiration = async () => {
+  private getStorageExpiration = async () => {
     const entryRes = await this.server.getLedgerEntries(
       new Contract(this.options.contractId).getFootprint()
     );
@@ -380,21 +594,6 @@ export class AssembledTransaction<T> {
     ];
   };
 
-  preImageFor(
-    entry: xdr.SorobanAuthorizationEntry,
-    signatureExpirationLedger: number
-  ): xdr.HashIdPreimage {
-    const addrAuth = entry.credentials().address();
-    return xdr.HashIdPreimage.envelopeTypeSorobanAuthorization(
-      new xdr.HashIdPreimageSorobanAuthorization({
-        networkId: hash(Buffer.from(this.options.networkPassphrase)),
-        nonce: addrAuth.nonce(),
-        invocation: entry.rootInvocation(),
-        signatureExpirationLedger,
-      })
-    );
-  }
-
   /**
    * If {@link needsNonInvokerSigningBy} returns a non-empty list, you can serialize
    * the transaction with `toJSON`, send it to the owner of one of the public keys
@@ -421,7 +620,7 @@ export class AssembledTransaction<T> {
      * contract's current `persistent` storage expiration date/ledger
      * number/block.
      */
-    expiration?: number | Promise<number> 
+    expiration?: number | Promise<number>
     /**
      * Sign all auth entries for this account. Default: the account that
      * constructed the transaction
@@ -491,6 +690,12 @@ export class AssembledTransaction<T> {
     }
   };
 
+  /**
+   * Whether this transaction is a read call. This is determined by the
+   * simulation result and the transaction data. If the transaction is a read
+   * call, it will not need to be signed and sent to the network. If this
+   * returns `false`, then you need to call `signAndSend` on this transaction.
+   */
   get isReadCall(): boolean {
     const authsCount = this.simulationData.result.auth.length;
     const writeLength = this.simulationData.transactionData
@@ -499,216 +704,4 @@ export class AssembledTransaction<T> {
       .readWrite().length;
     return authsCount === 0 && writeLength === 0;
   }
-}
-
-/**
- * A transaction that has been sent to the Soroban network. This happens in two steps:
- *
- * 1. `sendTransaction`: initial submission of the transaction to the network.
- *    This step can run into problems, and will be retried with exponential
- *    backoff if it does. See all attempts in `sendTransactionResponseAll` and the
- *    most recent attempt in `sendTransactionResponse`.
- * 2. `getTransaction`: once the transaction has been submitted to the network
- *    successfully, you need to wait for it to finalize to get the results of the
- *    transaction. This step can also run into problems, and will be retried with
- *    exponential backoff if it does. See all attempts in
- *    `getTransactionResponseAll` and the most recent attempt in
- *    `getTransactionResponse`.
- */
-class SentTransaction<T> {
-  public server: SorobanRpc.Server;
-  public signed?: Tx;
-  public sendTransactionResponse?: SendTx;
-  public getTransactionResponse?: GetTx;
-  public getTransactionResponseAll?: GetTx[];
-
-  static Errors = {
-    SendFailed: class SendFailedError extends Error {},
-    SendResultOnly: class SendResultOnlyError extends Error {},
-  }
-
-  constructor(
-    public signTransaction: ContractClientOptions["signTransaction"],
-    public assembled: AssembledTransaction<T>
-  ) {
-    if (!signTransaction) {
-      throw new Error(
-        "You must provide a `signTransaction` function to send a transaction"
-      );
-    }
-    this.server = new SorobanRpc.Server(this.assembled.options.rpcUrl, {
-      allowHttp: this.assembled.options.rpcUrl.startsWith("http://"),
-    });
-  }
-
-  static init = async <T>(
-    signTransaction: ContractClientOptions["signTransaction"],
-    assembled: AssembledTransaction<T>,
-  ): Promise<SentTransaction<T>> => {
-    const tx = new SentTransaction(signTransaction, assembled);
-    return await tx.send();
-  };
-
-  private send = async (): Promise<this> => {
-    const timeoutInSeconds = this.assembled.options.timeoutInSeconds ?? DEFAULT_TIMEOUT
-
-    const signature = await this.signTransaction!(
-      // `signAndSend` checks for `this.built` before calling `SentTransaction.init`
-      this.assembled.built!.toXDR(),
-      {
-        networkPassphrase: this.assembled.options.networkPassphrase,
-      }
-    );
-
-    this.signed = TransactionBuilder.fromXDR(
-      signature,
-      this.assembled.options.networkPassphrase
-    ) as Tx;
-
-    this.sendTransactionResponse = await this.server.sendTransaction(this.signed);
-
-    if (this.sendTransactionResponse.status !== "PENDING") {
-      throw new SentTransaction.Errors.SendFailed(
-        'Sending the transaction to the network failed!\n' +
-          JSON.stringify(
-            this.sendTransactionResponse,
-            null,
-            2
-          )
-      );
-    }
-
-    const { hash } = this.sendTransactionResponse;
-
-    this.getTransactionResponseAll = await withExponentialBackoff(
-      () => this.server.getTransaction(hash),
-      (resp) => resp.status === SorobanRpc.Api.GetTransactionStatus.NOT_FOUND,
-      timeoutInSeconds
-    );
-
-    this.getTransactionResponse =
-      this.getTransactionResponseAll[this.getTransactionResponseAll.length - 1];
-    if (
-      this.getTransactionResponse.status ===
-      SorobanRpc.Api.GetTransactionStatus.NOT_FOUND
-    ) {
-      console.error(
-        `Waited ${timeoutInSeconds} seconds for transaction to complete, but it did not. ` +
-          `Returning anyway. Check the transaction status manually. ` +
-          `Sent transaction: ${JSON.stringify(
-            this.sendTransactionResponse,
-            null,
-            2
-          )}\n` +
-          `All attempts to get the result: ${JSON.stringify(
-            this.getTransactionResponseAll,
-            null,
-            2
-          )}`
-      );
-    }
-
-    return this;
-  };
-
-  get result(): T {
-    // 1. check if transaction was submitted and awaited with `getTransaction`
-    if ("getTransactionResponse" in this && this.getTransactionResponse) {
-      // getTransactionResponse has a `returnValue` field unless it failed
-      if ("returnValue" in this.getTransactionResponse) {
-        return this.assembled.options.parseResultXdr(
-          this.getTransactionResponse.returnValue!
-        );
-      }
-
-      // if "returnValue" not present, the transaction failed; return without parsing the result
-      throw new Error("Transaction failed! Cannot parse result.");
-    }
-
-    // 2. otherwise, maybe it was merely sent with `sendTransaction`
-    if (this.sendTransactionResponse) {
-      const errorResult = this.sendTransactionResponse.errorResult?.result();
-      if (errorResult) {
-        throw new SentTransaction.Errors.SendFailed(
-          `Transaction simulation looked correct, but attempting to send the transaction failed. Check \`simulation\` and \`sendTransactionResponseAll\` to troubleshoot. Decoded \`sendTransactionResponse.errorResultXdr\`: ${errorResult}`
-        );
-      }
-      throw new SentTransaction.Errors.SendResultOnly(
-        `Transaction was sent to the network, but not yet awaited. No result to show. Await transaction completion with \`getTransaction(sendTransactionResponse.hash)\``
-      );
-    }
-
-    // 3. finally, if neither of those are present, throw an error
-    throw new Error(
-      `Sending transaction failed: ${JSON.stringify(this.assembled)}`
-    );
-  }
-}
-
-/**
- * Keep calling a `fn` for `timeoutInSeconds` seconds, if `keepWaitingIf` is true.
- * Returns an array of all attempts to call the function.
- */
-async function withExponentialBackoff<T>(
-  fn: (previousFailure?: T) => Promise<T>,
-  keepWaitingIf: (result: T) => boolean,
-  timeoutInSeconds: number,
-  exponentialFactor = 1.5,
-  verbose = false
-): Promise<T[]> {
-  const attempts: T[] = [];
-
-  let count = 0;
-  attempts.push(await fn());
-  if (!keepWaitingIf(attempts[attempts.length - 1])) return attempts;
-
-  const waitUntil = new Date(Date.now() + timeoutInSeconds * 1000).valueOf();
-  let waitTime = 1000;
-  let totalWaitTime = waitTime;
-
-  while (
-    Date.now() < waitUntil &&
-    keepWaitingIf(attempts[attempts.length - 1])
-  ) {
-    count++;
-    // Wait a beat
-    if (verbose) {
-      console.info(
-        `Waiting ${waitTime}ms before trying again (bringing the total wait time to ${totalWaitTime}ms so far, of total ${
-          timeoutInSeconds * 1000
-        }ms)`
-      );
-    }
-    await new Promise((res) => setTimeout(res, waitTime));
-    // Exponential backoff
-    waitTime = waitTime * exponentialFactor;
-    if (new Date(Date.now() + waitTime).valueOf() > waitUntil) {
-      waitTime = waitUntil - Date.now();
-      if (verbose) {
-        console.info(`was gonna wait too long; new waitTime: ${waitTime}ms`);
-      }
-    }
-    totalWaitTime = waitTime + totalWaitTime;
-    // Try again
-    attempts.push(await fn(attempts[attempts.length - 1]));
-    if (verbose && keepWaitingIf(attempts[attempts.length - 1])) {
-      console.info(
-        `${count}. Called ${fn}; ${
-          attempts.length
-        } prev attempts. Most recent: ${JSON.stringify(
-          attempts[attempts.length - 1],
-          null,
-          2
-        )}`
-      );
-    }
-  }
-
-  return attempts;
-}
-
-const contractErrorPattern = /Error\(Contract, #(\d+)\)/;
-
-function implementsToString(obj: unknown): obj is { toString(): string } {
-  return typeof obj === "object" && obj !== null && "toString" in obj;
 }
