+++ src/contract_client/index.ts
@@ -1,4 +1,6 @@
 export * from './assembled_transaction'
 export * from './basic_node_signer'
 export * from './client'
+export * from './sent_transaction'
 export * from './types'
+export * from './utils'
